requirements.txt
discord.py==1.1.1
tabulate==0.8.3
